Start the work on translations
We use the `tap:i18n` to define translation files for each supported
language. This commit extracts most of the English propositions in a
`en.i18n.json` file. From there it should be easy to translate the
application in many languages.

The following part are still hard-coded in English:
* Popups title (defined in the JavaScript code), but I plan to fix
  that with an upcoming Popup system refactor.
* Some of the activity messages. It's a bit difficult to mix HTML tags
  (like links) with translated propositions. Maybe the translation of
  activity messages should be handle with its own logic in the
  JavaScript code.
* Probably some attributes or texts I missed, but we will see them as
  soon as the first foreign translation is available.