Reduce the number of queries done in the "board" publication
Use the mongodb $or operator to express our conditions more
efficiently. There is still some work to do in this regard, especially
the subscription should be more granular and not publish every single
activity and comment just to display the board global view.

This commit also includes raw `publish` subscriptions for "card" and
"profile", just because we don't use `publishComposite` currently.