Make the new popup system reactive
The popup will be automatically repositioned if the window is resized.
The title will also change if the user change the page locale. We also
re-render the popup template less often because we use Tracker data
invalidation instead of removing from the DOM and adding it back with
new data.