Create a schema for Boards, Lists, Cards, and CardComments
This schema validation was really needed, first to ensure correctness
and avoid spelling bugs and other related issues, and second and
foremost because all the updates are done using minimongo
synchronization from the client. Which mean that the client could
literally post any document he wanted, adding new fields, etc.

Along with the new schemas this commit provide a migration package.
Until now we migrated the data by using the `meteor mongo` remote
shell, but this is not something really good, and anyway it was not
possible to do it for instances installed by third parties on their
own server (including the sandstorm platform).