Partially revert a643c80c
Collection2 autoValues break meteor latency compensation mechanism. So
we are back using collection-hooks to maintain createdAt and updatedAt
fields. The relevant collection2 bug/feature is available here:

  https://github.com/aldeed/meteor-collection2/issues/62

I'm still not sure why the before.insert hook is not called for lists
that are inserted from the server after the user creation. For now,
we duplicate the logic from the hook to the insert function, so it
doesn't block new user registration.

Fixes #102.